refactoring: move selectors and constants of users duck to the separated files
