+++ src/models/users/constants.js
@@ -1,0 +1,1 @@
+export const branch = 'users';
